+++ include/Graficador.h
@@ -17,23 +17,23 @@ class Graficador
 		void pintarEjes();
 		void pintarTitle(wxString strT);
 		void pintarLabels(wxString strH, wxString strV);
-		void pintarTicks(float mnx, float mxx, float mny, float mxy);
+		void pintarTicks(double mnx, double mxx, double mny, double mxy);
 
 		void pintarVariable(Variable *Var,int numConjunto,bool flagColor=true);
 		void pintarDifusor(Variable *Var, bool flagEjes=true);
 		void pintarConjunto(Variable *Var,ConjuntoDifuso *C);
 		void pintarPuntosArrastre(Variable *Var,ConjuntoDifuso *C);
 
-		void pintarFondoFuncion(float mnx, float mxx, float mny, float mxy, wxString nomEntra, wxString nomSale);
-		void pintarCurvaFuncion(float mnx, float mxx, float mny, float mxy,int numSalidas, int numEntra, int numSale, float* entra,wxColour &color);
+		void pintarFondoFuncion(double mnx, double mxx, double mny, double mxy, wxString nomEntra, wxString nomSale);
+		void pintarCurvaFuncion(double mnx, double mxx, double mny, double mxy,int numSalidas, int numEntra, int numSale, double* entra,wxColour &color);
 
 		void pintarFondoFuncionES(SistemaLogicaDifusa* SLD, int numEntra, int numSale);
-		void pintarCurvaFuncionES(SistemaLogicaDifusa* SLD, int numEntra, int numSale, float* entra,wxColour &color);
-		void pintarCurvaFuncionES(red* Red, int numEntra, int numSale, float* entra,wxColour &color);
+		void pintarCurvaFuncionES(SistemaLogicaDifusa* SLD, int numEntra, int numSale, double* entra,wxColour &color);
+		void pintarCurvaFuncionES(red* Red, int numEntra, int numSale, double* entra,wxColour &color);
 		void pintarComposicion(SistemaLogicaDifusa* SLD, int numSale, int numRegla);
-		void pintarConjuncion(SistemaLogicaDifusa* SLD, int numSale,float salida);
+		void pintarConjuncion(SistemaLogicaDifusa* SLD, int numSale,double salida);
 
-		void calcularFuncion(float*entra, float* sale);
+		void calcularFuncion(double*entra, double* sale);
 
 		int numeroPuntosArrastre;
 		wxPoint puntosArrastre[10];
@@ -55,7 +55,7 @@ class Graficador
 	  int numDX;
   	int numDY;
 
-		float arrastre[10];
+		double arrastre[10];
 	private:
 };
